experimental serch input: Add back token hover information
This PR extracts the token hover information extension into its own
module (including styles) to make it reusable and includes it in the
experimental search input.
