+++ client/branded/src/search-ui/input/CodeMirrorQueryInput.module.scss
@@ -66,41 +66,6 @@
             word-break: break-all;
         }
 
-        :global(.cm-tooltip) {
-            padding: 0.25rem;
-            color: var(--search-query-text-color);
-            background-color: var(--color-bg-1);
-            border: 1px solid var(--border-color);
-            border-radius: var(--border-radius);
-            box-shadow: var(--box-shadow);
-            max-width: 50vw;
-
-            p:last-child {
-                margin-bottom: 0;
-            }
-
-            code {
-                background-color: rgba(220, 220, 220, 0.4);
-                border-radius: var(--border-radius);
-                /* stylelint-disable-next-line declaration-property-unit-allowed-list */
-                padding: 0 0.4em;
-            }
-
-            :global(.cm-tooltip-section) {
-                padding-bottom: 0.25rem;
-                border-top-color: var(--border-color);
-
-                &:last-child {
-                    padding-top: 0.25rem;
-                    padding-bottom: 0;
-                }
-
-                &:last-child:first-child {
-                    padding: 0;
-                }
-            }
-        }
-
         :global(.cm-tooltip-autocomplete) {
             /* Resets padding added above to .cm-tooltip */
             padding: 0;
@@ -171,16 +136,17 @@
             }
         }
     }
-}
 
-// .placeholder needs to explicilty have the same background color because it
-// appears to be placed outside of .focusedFilter rather than within it.
-.placeholder,
-.focusedFilter {
-    background-color: var(--gray-02);
+    // .placeholder needs to explicilty have the same background color because it
+    // appears to be placed outside of .focusedFilter rather than within it.
+    .placeholder,
+    .focusedFilter,
+    :global(.sg-token-hover) {
+        background-color: var(--gray-02);
 
-    :global(.theme-dark) & {
-        background-color: var(--gray-08);
+        :global(.theme-dark) & {
+            background-color: var(--gray-08);
+        }
     }
 }
